requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
Pillow>=9.0.0
kaggle>=1.5.12
tqdm>=4.64.0
seaborn>=0.12.0
